refactor: move schema to server folder
