+++ src/app/[userId]/_actions.ts
@@ -1,6 +1,6 @@
 "use server";
 
-import { profilesTable } from "@/schema";
+import { profilesTable } from "@/server/schema";
 import { db } from "@/server/db";
 import { eq } from "drizzle-orm";
 import { revalidatePath } from "next/cache";
